+++ backend/src/main/java/com/linyi/pig/service/impl/ConversationServiceImpl.java
@@ -74,8 +74,8 @@ public class ConversationServiceImpl extends ServiceImpl<ConversationMapper, Con
         Page<Conversation> page = new Page<>(conversationQueryVo.getPageNum(),
                 conversationQueryVo.getPageSize());
         Page<Conversation> pageNew = conversationMapper.selectPage(page, queryWrapper);
-        return new PageResult<>(pageNew.getRecords(), pageNew.getTotal(),
-                conversationQueryVo.getPageNum(), conversationQueryVo.getPageSize(), pageNew.getPages());
+        return new PageResult<>(pageNew.getRecords(), (int)pageNew.getTotal(),
+                conversationQueryVo.getPageNum(), conversationQueryVo.getPageSize(), (int)pageNew.getPages());
     }
 
     @Override
