+++ backend/src/main/java/com/linyi/pig/service/impl/FilesServiceImpl.java
@@ -72,7 +72,7 @@ public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files> implements
         //查询数据
         Page<Files> pageNew = filesMapper.selectPage(page, queryWrapper);
         //返回分页数据
-        return new PageResult<>(pageNew.getRecords(), pageNew.getTotal(), filesQueryVo.getPageNum(), filesQueryVo.getPageSize(), pageNew.getPages());
+        return new PageResult<>(pageNew.getRecords(), (int)pageNew.getTotal(), filesQueryVo.getPageNum(), filesQueryVo.getPageSize(), (int)pageNew.getPages());
     }
 
     @Override
